+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-01 03:46:59
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 891e4ac29cc20a9b23173543a08f8f0f78e12089
+
+## [ForwardWidgets-2kuai] - 2025-08-01 03:46:59
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 5d49f3d57c09ba4f809fa15fc6aa84e8cdf20570
+
+## [ForwardWidgets-pack1r] - 2025-08-01 03:46:59
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: b4fd6b63d3dd52576f661d9f4cd4f347e5d5d0c2
+
+
 ## [ForwardWidgets-huangxd] - 2025-07-31 03:25:10
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main